+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+
+0.14.1 (2023-04-26)
+===================
+
+* Hot fix release, loosing pin requirements for `barril`.
+
 0.14.0 (2022-07-19)
 ===================